mobilenetv3ssd/requirements.txt
absl-py==1.1.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
chitra==0.2.0 				#for preprocessing only
click==8.1.3
cycler==0.11.0
ensemble-boxes==1.0.9 		#confusion matrix
fonttools==4.34.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.21.0
imgaug==0.4.0
importlib-metadata==4.12.0
kiwisolver==1.4.3
llvmlite==0.39.1
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.6.3
numba==0.56.2
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.0 			#confusion matrix
Shapely==1.8.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
torch==1.12.0
torchmetrics==0.9.3
torchvision==0.13.0
typeguard==2.13.3
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
